refactor(transactions): decompose TransactionDetailView into per-section subviews
Splits the 1213-line TransactionDetailView (875-line struct body) into a
thin parent view plus 19 focused subviews under
`Features/Transactions/Views/Detail/`. The parent is now 392 lines with a
147-line struct body, well under the SwiftLint thresholds (400/250). The
`type_body_length` and `file_length` suppression comments are gone, and
`.swiftlint-baseline.yml` is unchanged.

Notable structural changes:

- Replaces three parallel `[Int: ...]` dicts (`showLegCategorySuggestions`,
  `legCategoryHighlightedIndex`, `legCategoryJustSelected`) with a single
  `[Int: CategoryAutocompleteState]` dict, and moves their reset paths
  onto `CategoryAutocompleteState.dismiss()` / `PayeeAutocompleteState
  .dismiss()` so the "close dropdown without re-firing onTextChange"
  invariant lives in one place.
- Drops the parent's cross-leg `.onChange(of: legCategoryFieldFocused)`
  handler. Each `TransactionDetailLegRow` now owns its own `@FocusState`
  for its category field and runs its own blur cleanup; SwiftUI's
  exclusive-focus guarantee means this fires exactly when needed.
- Re-keys the per-leg dropdown state dict when a leg is deleted so the
  state previously associated with leg `n+1` doesn't leak onto the new
  leg `n` after deletion (`shiftLegCategoryStates(after:)`).
- Adds `TransactionDraft.normaliseLegCategoryText(at:using:)` so the
  per-leg blur handler delegates text-vs-id reconciliation to the draft
  (testable directly, no view host required).
- Promotes the focus enum to `TransactionDetailFocus` so subviews can
  drive `@FocusState.Binding` for `.legAmount(index)` etc. without owning
  the enum themselves.

Closes #309
